chore(refactor): split out schema types
